Finished calendar entry fields extraction. It required to add code to manually extract the XML attributes from a xmlNode (I don't known how to extract the value directly from XML_ATTRIBUTE_NODE).
As a result, when calling 'extract_and_check' function, I need to supply
an extra parameter for the attribute name (NULL if the XPath expression
already returns a XML_TEXT_NODE).

To extract the event data of all entries, is just a matter of looping over
the nodes and call 'atom_extract_data' (I think it will be done easily).