Adding an extra parameter to make google server return ordered contacts by 'last-updated'.
Google contacts behavior is different from calendar: the most recent updated
entry is returned as last element (calendar is the contrary: the first element
is the most recently updated event).

This is a requirement to make possible to store a timestamp to be used to
do fast-sync (where only the updated elements are returned).